requirements.txt
apache-airflow
pandas
requests
boto3
python-dotenv
snowflake-connector-python